Moved litter enviromental factors into a seperate module
