+++ virtual_ecosystem/models/litter/env_factors.py
@@ -1,0 +1,129 @@
+"""The ``models.litter.env_factors`` module contains functions that are used to
+capture the impact that environmental factors have on litter decay rates. These include
+temperature and soil water potential.
+"""  # noqa: D205
+
+import numpy as np
+from numpy.typing import NDArray
+
+from virtual_ecosystem.models.litter.constants import LitterConsts
+
+
+def calculate_environmental_factors(
+    surface_temp: NDArray[np.float32],
+    topsoil_temp: NDArray[np.float32],
+    water_potential: NDArray[np.float32],
+    constants: LitterConsts,
+) -> dict[str, NDArray[np.float32]]:
+    """Calculate the impact of the environment has on litter decay across litter layers.
+
+    For the above ground layer the impact of temperature is calculated, and for the
+    below ground layer the effect of temperature and soil water potential are
+    considered.
+
+    Args:
+        surface_temp: Temperature of soil surface, which is assumed to be the same
+            temperature as the above ground litter [C]
+        topsoil_temp: Temperature of topsoil layer, which is assumed to be the same
+            temperature as the below ground litter [C]
+        water_potential: Water potential of the topsoil layer [kPa]
+        constants: Set of constants for the litter model
+
+    Returns:
+        A dictionary containing three environmental factors, one for the effect of
+        temperature on above ground litter decay, one for the effect of temperature on
+        below ground litter decay, and one for the effect of soil water potential on
+        below ground litter decay.
+    """
+    # Calculate temperature factor for the above ground litter layers
+    temperature_factor_above = calculate_temperature_effect_on_litter_decomp(
+        temperature=surface_temp,
+        reference_temp=constants.litter_decomp_reference_temp,
+        offset_temp=constants.litter_decomp_offset_temp,
+        temp_response=constants.litter_decomp_temp_response,
+    )
+    # Calculate temperature factor for the below ground litter layers
+    temperature_factor_below = calculate_temperature_effect_on_litter_decomp(
+        temperature=topsoil_temp,
+        reference_temp=constants.litter_decomp_reference_temp,
+        offset_temp=constants.litter_decomp_offset_temp,
+        temp_response=constants.litter_decomp_temp_response,
+    )
+    # Calculate the water factor (relevant for below ground layers)
+    water_factor = calculate_soil_water_effect_on_litter_decomp(
+        water_potential=water_potential,
+        water_potential_halt=constants.litter_decay_water_potential_halt,
+        water_potential_opt=constants.litter_decay_water_potential_optimum,
+        moisture_response_curvature=constants.moisture_response_curvature,
+    )
+
+    # Return all three factors in a single dictionary
+    return {
+        "temp_above": temperature_factor_above,
+        "temp_below": temperature_factor_below,
+        "water": water_factor,
+    }
+
+
+def calculate_temperature_effect_on_litter_decomp(
+    temperature: NDArray[np.float32],
+    reference_temp: float,
+    offset_temp: float,
+    temp_response: float,
+) -> NDArray[np.float32]:
+    """Calculate the effect that temperature has on litter decomposition rates.
+
+    This function is taken from :cite:t:`kirschbaum_modelling_2002`.
+
+    Args:
+        temperature: The temperature of the litter layer [C]
+        reference_temp: The reference temperature for changes in litter decomposition
+            rates with temperature [C]
+        offset_temp: Temperature offset [C]
+        temp_response: Factor controlling response strength to changing temperature
+            [unitless]
+
+    Returns:
+        A multiplicative factor capturing the impact of temperature on litter
+        decomposition [unitless]
+    """
+
+    return np.exp(
+        temp_response * (temperature - reference_temp) / (temperature + offset_temp)
+    )
+
+
+def calculate_soil_water_effect_on_litter_decomp(
+    water_potential: NDArray[np.float32],
+    water_potential_halt: float,
+    water_potential_opt: float,
+    moisture_response_curvature: float,
+) -> NDArray[np.float32]:
+    """Calculate the effect that soil water potential has on litter decomposition rates.
+
+    This function is only relevant for the below ground litter pools. Its functional
+    form is taken from :cite:t:`moyano_responses_2013`.
+
+    Args:
+        water_potential: Soil water potential [kPa]
+        water_potential_halt: Water potential at which all microbial activity stops
+            [kPa]
+        water_potential_opt: Optimal water potential for microbial activity [kPa]
+        moisture_response_curvature: Parameter controlling the curvature of the moisture
+            response function [unitless]
+
+    Returns:
+        A multiplicative factor capturing the impact of moisture on below ground litter
+        decomposition [unitless]
+    """
+
+    # TODO - Need to make sure that this function is properly defined for a plausible
+    # range of matric potentials.
+
+    # Calculate how much moisture suppresses microbial activity
+    supression = (
+        (np.log10(-water_potential) - np.log10(-water_potential_opt))
+        / (np.log10(-water_potential_halt) - np.log10(-water_potential_opt))
+    ) ** moisture_response_curvature
+
+    return 1 - supression
